--- src/api/character/routes/01-custom-routes.js
@@ -1,9 +1,0 @@
-module.exports = {
-  routes: [
-    { // Path defined with an URL parameter
-      method: 'POST',
-      path: '/characters/purchased/:paymentId', 
-      handler: 'character.characterPurchased',
-    }
-  ]
-}
